src/prompt_augmentation/base.py
from abc import ABC, abstractmethod
from typing import Dict, List, Optional, Union
import openai
from dotenv import load_dotenv

load_dotenv()

class BasePromptAugmenter(ABC):
    """Base class for prompt augmentation strategies."""
    
    def __init__(self, model: str = "gpt-3.5-turbo"):
        """
        Initialize the prompt augmenter.
        
        Args:
            model (str): The OpenAI model to use for augmentation
        """
        self.model = model
        self.client = openai.OpenAI()
    
    @abstractmethod
    def augment_prompt(self, prompt: str, **kwargs) -> str:
        """
        Augment the given prompt using the specific strategy.
        
        Args:
            prompt (str): The original prompt to augment
            **kwargs: Additional arguments specific to the augmentation strategy
            
        Returns:
            str: The augmented prompt
        """
        pass
    
    def generate_response(self, prompt: str, **kwargs) -> str:
        """
        Generate a response using the OpenAI API.
        
        Args:
            prompt (str): The prompt to generate a response for
            **kwargs: Additional arguments for the API call
            
        Returns:
            str: The generated response
        """
        response = self.client.chat.completions.create(
            model=self.model,
            messages=[{"role": "user", "content": prompt}],
            **kwargs
        )
        return response.choices[0].message.content
    
    def evaluate_alignment(self, response: str, criteria: List[str]) -> Dict[str, float]:
        """
        Evaluate the alignment of a response with given criteria.
        
        Args:
            response (str): The response to evaluate
            criteria (List[str]): List of alignment criteria to check
            
        Returns:
            Dict[str, float]: Scores for each criterion
        """
        # TODO: Implement alignment evaluation logic
        return {criterion: 0.0 for criterion in criteria} 