requirements.txt
openai>=1.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
pytest>=6.2.5
jupyter>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.62.0
matplotlib>=3.4.3
seaborn>=0.11.2
scipy>=1.7.0
plotly>=5.0.0
textstat>=0.7.0
argparse 